# Optimizing "grafana generated" regex matchers