When using a multi-value variable on Grafana Dashboards the default behavior is to produce regex in the format `(a|b)`. See https://grafana.com/docs/grafana/latest/dashboards/variables/variable-syntax/#regex

This PR is expanding the optimization Prometheus already have on regexes like `a|b` to regexes like `(a|b)` and improve the query time and cpu utilization for dashboards with high cardinality variables substantially.

PS: Nothing will change for other types of queries even if they starts and ends with `()`, like  `(a|b)|(c|d)`

```
goos: linux
goarch: amd64
pkg: github.com/prometheus/prometheus/tsdb
cpu: Intel(R) Xeon(R) Platinum 8259CL CPU @ 2.50GHz
                                                             │     /tmp/old     │              /tmp/new               │
                                                             │      sec/op      │    sec/op     vs base               │
Querier/Head/PostingsForMatchers/i=~"(1|2|3|4|5|6|20|55)"-32   18796.184µ ± ∞ ¹   2.416µ ± ∞ ¹  -99.99% (p=0.008 n=5)
Querier/Head/PostingsForMatchers/i!~"(1|2|3|4|5|6|20|55)"-32      18.450m ± ∞ ¹   2.238m ± ∞ ¹  -87.87% (p=0.008 n=5)
geomean                                                            18.62m         73.53µ        -99.61%
¹ need >= 6 samples for confidence interval at level 0.95

                                                             │    /tmp/old     │               /tmp/new               │
                                                             │      B/op       │     B/op	vs base               │
Querier/Head/PostingsForMatchers/i=~"(1|2|3|4|5|6|20|55)"-32   1606349.0 ± ∞ ¹     848.0 ± ∞ ¹  -99.95% (p=0.008 n=5)
Querier/Head/PostingsForMatchers/i!~"(1|2|3|4|5|6|20|55)"-32     1.531Mi ± ∞ ¹   1.532Mi ± ∞ ¹   +0.04% (p=0.008 n=5)
geomean                                                          1.532Mi         36.04Ki        -97.70%
¹ need >= 6 samples for confidence interval at level 0.95

                                                             │  /tmp/old   │               /tmp/new               │
                                                             │  allocs/op  │  allocs/op    vs base                │
Querier/Head/PostingsForMatchers/i=~"(1|2|3|4|5|6|20|55)"-32   4.000 ± ∞ ¹   31.000 ± ∞ ¹  +675.00% (p=0.008 n=5)
Querier/Head/PostingsForMatchers/i!~"(1|2|3|4|5|6|20|55)"-32   7.000 ± ∞ ¹   35.000 ± ∞ ¹  +400.00% (p=0.008 n=5)
geomean                                                        5.292          32.94        +522.49%
```